fix: simplify errors and pass tests

+++ apps/server/src/services/usuarios.service.ts
@@ -1,18 +1,26 @@
 import type { Usuario } from "@prisma/client";
+import { ConflictError } from "@server/exceptions/ConflictError";
+import { NotFoundError } from "@server/exceptions/NotFoundError";
+import { ValidationError } from "@server/exceptions/ValidationError";
+import { EventosRepository } from "@server/repositories/eventos.repository";
+import { InscripcionesRepository } from "@server/repositories/inscripciones.repository";
+import { UsuariosRepository } from "@server/repositories/usuarios.repository";
+import type { CreateUsuarioInput } from "@server/schemas/usuarios/create-usuario.schema";
+import type { UpdateUsuarioInput } from "@server/schemas/usuarios/update-usuario.schema";
 import bcrypt from "bcryptjs";
-import { ConflictError } from "@/exceptions/ConflictError";
-import { NotFoundError } from "@/exceptions/NotFoundError";
-import { ValidationError } from "@/exceptions/ValidationError";
-import { InscripcionesRepository } from "@/repositories/inscripciones.repository";
-import { UsuariosRepository } from "@/repositories/usuarios.repository";
-import type { CreateUsuarioInput } from "@/schemas/usuarios/create-usuario.schema";
-import type { UpdateUsuarioInput } from "@/schemas/usuarios/update-usuario.schema";
 
 // Servicio para manejar la lógica de negocio relacionada con usuarios
 
 export const UsuariosService = {
   async delete(id: string) {
+    const eventos = await EventosRepository.findByUserId(id);
+    if (eventos.length > 0) {
+      throw new ConflictError("Usuario tiene eventos asociados");
+    }
+
     const eliminado = await UsuariosRepository.delete(id);
+    console.log("eliminado", eliminado);
+
     if (!eliminado) {
       throw new NotFoundError("Usuario no encontrado");
     }
